requirements.txt
# cat requirements.txt 
setuptools>=57.0.0
pyspark>=3.1.2
pygeohash>=1.2.0
requests>=2.25.1
python-dotenv>=0.20.0